+++ defaults-themes.php
@@ -1,0 +1,101 @@
+<?php
+function custom_admin_styles() {
+  ?>
+  <style>
+/*   TODO: Uncomment by Brent */
+/*     .cf-select__input option[value=nad-denim] { */
+/*       display: none !important; */
+/*     } */
+
+/*     .cf-select__input option[value=nad-nile] { */
+/*       display: none !important; */
+/*     } */
+
+/*     .cf-select__input option[value=nad-amethyst] { */
+/*       display: none !important; */
+/*     } */
+
+/*     .cf-select__input option[value=nad-spark] { */
+/*       display: none !important; */
+/*     } */
+
+/*     .cf-select__input option[value=nad-miracle] { */
+/*       display: none !important; */
+/*     } */
+
+/*     .cf-select__input option[value=nad-branch] { */
+/*       display: none !important; */
+/*     } */
+
+/*     .cf-select__input option[value=nad-vine] { */
+/*       display: none !important; */
+/*     } */
+
+    /* .cf-select__input option[value=treefrog] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=ming] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=bluejay] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=iris] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=lily] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=scarlett] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=campfire] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=winter] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=forest] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=cave] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=denim] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=emperor] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=grapevine] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=velvet] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=earth] {
+      display: none !important;
+    } */
+
+    /* .cf-select__input option[value=night] {
+      display: none !important;
+    } */
+
+  </style>
+  <?php
+}
+add_action('admin_head', 'custom_admin_styles');